server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]
pydantic-settings==2.0.3
python-dotenv==1.1.0

asyncpg==0.29.0
databases[postgresql]==0.8.0
sqlalchemy[asyncio]==1.4.49

bcrypt==4.1.2
PyJWT==2.8.0

python-multipart==0.0.6
PyMuPDF==1.23.9
aiofiles==23.2.1

numpy==2.2.5
sentence-transformers==4.1.0
pymilvus==2.5.7

motor==3.3.2
pymongo==4.6.0
beanie==1.24.0


